+++ spark/src/main/java/edu/berkeley/cs/succinct/streams/SuccinctStream.java
@@ -296,7 +296,7 @@ public class SuccinctStream extends SuccinctCore {
    * @param i Index into inverted alphabet map
    * @return Value of inverted alphabet map at specified index.
    */
-  @Override public long lookupC(long i) {
+  @Override public int lookupC(long i) {
     try {
       return SerializedOperations.ArrayOps.getRank1(coloffsets, 0, getSigmaSize(), i) - 1;
     } catch (IOException e) {
